+++ api.tsx
@@ -34,7 +34,6 @@ export const visibleReset = () => {
 };
 export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
   const [_, option, value, town, city] = queryKey;
-
   const getData: { [key: string]: string }[] = [];
   let q;
 
@@ -47,7 +46,7 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
         orderBy(option),
         startAt(value),
         endAt(value + '\uf8ff'),
-        limit(4),
+        limit(8),
         startAfter(lastVisible)
       );
     } else if (value) {
@@ -56,7 +55,7 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
         orderBy(option),
         startAt(value),
         endAt(value + '\uf8ff'),
-        limit(8)
+        limit(20)
       );
     } else {
       if (town && lastVisible) {
@@ -64,7 +63,7 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
           collection(dbService, 'post'),
           where('town', '==', town),
           orderBy('createdAt', 'desc'),
-          limit(4),
+          limit(8),
           startAfter(lastVisible)
         );
       } else if (town) {
@@ -72,7 +71,7 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
           collection(dbService, 'post'),
           where('town', '==', town),
           orderBy('createdAt', 'desc'),
-          limit(8)
+          limit(20)
         );
       } else {
         if (city && lastVisible) {
@@ -80,7 +79,7 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
             collection(dbService, 'post'),
             where('city', '==', city),
             orderBy('createdAt', 'desc'),
-            limit(4),
+            limit(8),
             startAfter(lastVisible)
           );
         } else if (city) {
@@ -88,21 +87,21 @@ export const getInfiniteData = async ({ queryKey }: { queryKey: string[] }) => {
             collection(dbService, 'post'),
             where('city', '==', city),
             orderBy('createdAt', 'desc'),
-            limit(8)
+            limit(20)
           );
         } else {
           if (lastVisible) {
             q = query(
               collection(dbService, 'post'),
               orderBy('createdAt', 'desc'),
-              limit(4),
+              limit(8),
               startAfter(lastVisible)
             );
           } else {
             q = query(
               collection(dbService, 'post'),
               orderBy('createdAt', 'desc'),
-              limit(8)
+              limit(20)
             );
           }
         }
